Move ckpt stuff
