+++ common/candle/__init__.py
@@ -87,9 +87,9 @@ if 'keras' in sys.modules:
     from keras_utils import PermanentDropout
     from keras_utils import register_permanent_dropout
     from keras_utils import LoggingCallback
-    from keras_utils import MultiGPUCheckpoint
-    from keras_utils import CandleCheckpointCallback
-    from keras_utils import restart
+    from ckpt_keras_utils import MultiGPUCheckpoint
+    from ckpt_keras_utils import CandleCheckpointCallback
+    from ckpt_keras_utils import restart
     from keras_utils import r2
     from keras_utils import mae
     from keras_utils import mse
